Struct aws_sdk_comprehend::client::fluent_builders::StartSentimentDetectionJob

pub struct StartSentimentDetectionJob<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Implementations

impl<C, M, R> StartSentimentDetectionJob<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(
self
) -> Result<StartSentimentDetectionJobOutput, SdkError<StartSentimentDetectionJobError>> where
R::Policy: SmithyRetryPolicy<StartSentimentDetectionJobInputOperationOutputAlias, StartSentimentDetectionJobOutput, StartSentimentDetectionJobError, StartSentimentDetectionJobInputOperationRetryAlias>,

pub fn input_data_config(self, inp: InputDataConfig) -> Self

Specifies the format and location of the input data for the job.

pub fn set_input_data_config(self, input: Option<InputDataConfig>) -> Self

pub fn output_data_config(self, inp: OutputDataConfig) -> Self

Specifies where to send the output files.

pub fn set_output_data_config(self, input: Option<OutputDataConfig>) -> Self

pub fn data_access_role_arn(self, inp: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the AWS Identity and Access Management (IAM) role that grants Amazon Comprehend read access to your input data. For more information, see https://docs.aws.amazon.com/comprehend/latest/dg/access-control-managing-permissions.html#auth-role-permissions.

pub fn set_data_access_role_arn(self, input: Option<String>) -> Self

pub fn job_name(self, inp: impl Into<String>) -> Self

The identifier of the job.

pub fn set_job_name(self, input: Option<String>) -> Self

pub fn language_code(self, inp: LanguageCode) -> Self

The language of the input documents. You can specify any of the primary languages supported by Amazon Comprehend. All documents must be in the same language.

pub fn set_language_code(self, input: Option<LanguageCode>) -> Self

pub fn client_request_token(self, inp: impl Into<String>) -> Self

A unique identifier for the request. If you don't set the client request token, Amazon Comprehend generates one.

pub fn set_client_request_token(self, input: Option<String>) -> Self

pub fn volume_kms_key_id(self, inp: impl Into<String>) -> Self

ID for the AWS Key Management Service (KMS) key that Amazon Comprehend uses to encrypt data on the storage volume attached to the ML compute instance(s) that process the analysis job. The VolumeKmsKeyId can be either of the following formats:

KMS Key ID: "1234abcd-12ab-34cd-56ef-1234567890ab"

Amazon Resource Name (ARN) of a KMS Key: "arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab"

pub fn set_volume_kms_key_id(self, input: Option<String>) -> Self

pub fn vpc_config(self, inp: VpcConfig) -> Self

Configuration parameters for an optional private Virtual Private Cloud (VPC) containing the resources you are using for your sentiment detection job. For more information, see Amazon VPC.

pub fn set_vpc_config(self, input: Option<VpcConfig>) -> Self

pub fn tags(self, inp: impl Into<Tag>) -> Self

Appends an item to Tags.

To override the contents of this collection use set_tags.

Tags to be associated with the sentiment detection job. A tag is a key-value pair that adds metadata to a resource used by Amazon Comprehend. For example, a tag with "Sales" as the key might be added to a resource to indicate its use by the sales department.

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self