Module aws_sdk_comprehend::error

Modules

batch_size_limit_exceeded_exception

See BatchSizeLimitExceededException

concurrent_modification_exception

See ConcurrentModificationException

internal_server_exception

See InternalServerException

invalid_filter_exception

See InvalidFilterException

invalid_request_exception

See InvalidRequestException

job_not_found_exception

See JobNotFoundException

kms_key_validation_exception

See KmsKeyValidationException

resource_in_use_exception

See ResourceInUseException

resource_limit_exceeded_exception

See ResourceLimitExceededException

resource_not_found_exception

See ResourceNotFoundException

resource_unavailable_exception

See ResourceUnavailableException

text_size_limit_exceeded_exception

See TextSizeLimitExceededException

too_many_requests_exception

See TooManyRequestsException

too_many_tag_keys_exception

See TooManyTagKeysException

too_many_tags_exception

See TooManyTagsException

unsupported_language_exception

See UnsupportedLanguageException

Structs

BatchDetectDominantLanguageError

BatchDetectEntitiesError

BatchDetectKeyPhrasesError

BatchDetectSentimentError

BatchDetectSyntaxError

BatchSizeLimitExceededException

The number of documents in the request exceeds the limit of 25. Try your request again

ClassifyDocumentError

ConcurrentModificationException

Concurrent modification of the tags associated with an Amazon Comprehend resource is not

ContainsPiiEntitiesError

CreateDocumentClassifierError

CreateEndpointError

CreateEntityRecognizerError

DeleteDocumentClassifierError

DeleteEndpointError

DeleteEntityRecognizerError

DescribeDocumentClassificationJobError

DescribeDocumentClassifierError

DescribeDominantLanguageDetectionJobError

DescribeEndpointError

DescribeEntitiesDetectionJobError

DescribeEntityRecognizerError

DescribeEventsDetectionJobError

DescribeKeyPhrasesDetectionJobError

DescribePiiEntitiesDetectionJobError

DescribeSentimentDetectionJobError

DescribeTopicsDetectionJobError

DetectDominantLanguageError

DetectEntitiesError

DetectKeyPhrasesError

DetectPiiEntitiesError

DetectSentimentError

DetectSyntaxError

InternalServerException

An internal server error occurred. Retry your request.

InvalidFilterException

The filter specified for the operation is invalid. Specify a different

InvalidRequestException

The request is invalid.

JobNotFoundException

The specified job was not found. Check the job ID and try again.

KmsKeyValidationException

The KMS customer managed key (CMK) entered cannot be validated. Verify the key and

ListDocumentClassificationJobsError

ListDocumentClassifiersError

ListDominantLanguageDetectionJobsError

ListEndpointsError

ListEntitiesDetectionJobsError

ListEntityRecognizersError

ListEventsDetectionJobsError

ListKeyPhrasesDetectionJobsError

ListPiiEntitiesDetectionJobsError

ListSentimentDetectionJobsError

ListTagsForResourceError

ListTopicsDetectionJobsError

ResourceInUseException

The specified resource name is already in use. Use a different name and try your request

ResourceLimitExceededException

The maximum number of resources per account has been exceeded. Review the resources, and

ResourceNotFoundException

The specified resource ARN was not found. Check the ARN and try your request again.

ResourceUnavailableException

The specified resource is not available. Check the resource and try your request

StartDocumentClassificationJobError

StartDominantLanguageDetectionJobError

StartEntitiesDetectionJobError

StartEventsDetectionJobError

StartKeyPhrasesDetectionJobError

StartPiiEntitiesDetectionJobError

StartSentimentDetectionJobError

StartTopicsDetectionJobError

StopDominantLanguageDetectionJobError

StopEntitiesDetectionJobError

StopEventsDetectionJobError

StopKeyPhrasesDetectionJobError

StopPiiEntitiesDetectionJobError

StopSentimentDetectionJobError

StopTrainingDocumentClassifierError

StopTrainingEntityRecognizerError

TagResourceError

TextSizeLimitExceededException

The size of the input text exceeds the limit. Use a smaller document.

TooManyRequestsException

The number of requests exceeds the limit. Resubmit your request later.

TooManyTagKeysException

The request contains more tag keys than can be associated with a resource (50 tag keys per

TooManyTagsException

The request contains more tags than can be associated with a resource (50 tags per

UnsupportedLanguageException

Amazon Comprehend can't process the language of the input text. For custom entity

UntagResourceError

UpdateEndpointError

Enums

BatchDetectDominantLanguageErrorKind

BatchDetectEntitiesErrorKind

BatchDetectKeyPhrasesErrorKind

BatchDetectSentimentErrorKind

BatchDetectSyntaxErrorKind

ClassifyDocumentErrorKind

ContainsPiiEntitiesErrorKind

CreateDocumentClassifierErrorKind

CreateEndpointErrorKind

CreateEntityRecognizerErrorKind

DeleteDocumentClassifierErrorKind

DeleteEndpointErrorKind

DeleteEntityRecognizerErrorKind

DescribeDocumentClassificationJobErrorKind

DescribeDocumentClassifierErrorKind

DescribeDominantLanguageDetectionJobErrorKind

DescribeEndpointErrorKind

DescribeEntitiesDetectionJobErrorKind

DescribeEntityRecognizerErrorKind

DescribeEventsDetectionJobErrorKind

DescribeKeyPhrasesDetectionJobErrorKind

DescribePiiEntitiesDetectionJobErrorKind

DescribeSentimentDetectionJobErrorKind

DescribeTopicsDetectionJobErrorKind

DetectDominantLanguageErrorKind

DetectEntitiesErrorKind

DetectKeyPhrasesErrorKind

DetectPiiEntitiesErrorKind

DetectSentimentErrorKind

DetectSyntaxErrorKind

ListDocumentClassificationJobsErrorKind

ListDocumentClassifiersErrorKind

ListDominantLanguageDetectionJobsErrorKind

ListEndpointsErrorKind

ListEntitiesDetectionJobsErrorKind

ListEntityRecognizersErrorKind

ListEventsDetectionJobsErrorKind

ListKeyPhrasesDetectionJobsErrorKind

ListPiiEntitiesDetectionJobsErrorKind

ListSentimentDetectionJobsErrorKind

ListTagsForResourceErrorKind

ListTopicsDetectionJobsErrorKind

StartDocumentClassificationJobErrorKind

StartDominantLanguageDetectionJobErrorKind

StartEntitiesDetectionJobErrorKind

StartEventsDetectionJobErrorKind

StartKeyPhrasesDetectionJobErrorKind

StartPiiEntitiesDetectionJobErrorKind

StartSentimentDetectionJobErrorKind

StartTopicsDetectionJobErrorKind

StopDominantLanguageDetectionJobErrorKind

StopEntitiesDetectionJobErrorKind

StopEventsDetectionJobErrorKind

StopKeyPhrasesDetectionJobErrorKind

StopPiiEntitiesDetectionJobErrorKind

StopSentimentDetectionJobErrorKind

StopTrainingDocumentClassifierErrorKind

StopTrainingEntityRecognizerErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateEndpointErrorKind