Module aws_sdk_comprehend::model

Modules

augmented_manifests_list_item

See AugmentedManifestsListItem

batch_detect_dominant_language_item_result

See BatchDetectDominantLanguageItemResult

batch_detect_entities_item_result

See BatchDetectEntitiesItemResult

batch_detect_key_phrases_item_result

See BatchDetectKeyPhrasesItemResult

batch_detect_sentiment_item_result

See BatchDetectSentimentItemResult

batch_detect_syntax_item_result

See BatchDetectSyntaxItemResult

batch_item_error

See BatchItemError

classifier_evaluation_metrics

See ClassifierEvaluationMetrics

classifier_metadata

See ClassifierMetadata

document_class

See DocumentClass

document_classification_job_filter

See DocumentClassificationJobFilter

document_classification_job_properties

See DocumentClassificationJobProperties

document_classifier_filter

See DocumentClassifierFilter

document_classifier_input_data_config

See DocumentClassifierInputDataConfig

document_classifier_output_data_config

See DocumentClassifierOutputDataConfig

document_classifier_properties

See DocumentClassifierProperties

document_label

See DocumentLabel

dominant_language

See DominantLanguage

dominant_language_detection_job_filter

See DominantLanguageDetectionJobFilter

dominant_language_detection_job_properties

See DominantLanguageDetectionJobProperties

endpoint_filter

See EndpointFilter

endpoint_properties

See EndpointProperties

entities_detection_job_filter

See EntitiesDetectionJobFilter

entities_detection_job_properties

See EntitiesDetectionJobProperties

entity

See Entity

entity_label

See EntityLabel

entity_recognizer_annotations

See EntityRecognizerAnnotations

entity_recognizer_documents

See EntityRecognizerDocuments

entity_recognizer_entity_list

See EntityRecognizerEntityList

entity_recognizer_evaluation_metrics

See EntityRecognizerEvaluationMetrics

entity_recognizer_filter

See EntityRecognizerFilter

entity_recognizer_input_data_config

See EntityRecognizerInputDataConfig

entity_recognizer_metadata

See EntityRecognizerMetadata

entity_recognizer_metadata_entity_types_list_item

See EntityRecognizerMetadataEntityTypesListItem

entity_recognizer_properties

See EntityRecognizerProperties

entity_types_evaluation_metrics

See EntityTypesEvaluationMetrics

entity_types_list_item

See EntityTypesListItem

events_detection_job_filter

See EventsDetectionJobFilter

events_detection_job_properties

See EventsDetectionJobProperties

input_data_config

See InputDataConfig

key_phrase

See KeyPhrase

key_phrases_detection_job_filter

See KeyPhrasesDetectionJobFilter

key_phrases_detection_job_properties

See KeyPhrasesDetectionJobProperties

output_data_config

See OutputDataConfig

part_of_speech_tag

See PartOfSpeechTag

pii_entities_detection_job_filter

See PiiEntitiesDetectionJobFilter

pii_entities_detection_job_properties

See PiiEntitiesDetectionJobProperties

pii_entity

See PiiEntity

pii_output_data_config

See PiiOutputDataConfig

redaction_config

See RedactionConfig

sentiment_detection_job_filter

See SentimentDetectionJobFilter

sentiment_detection_job_properties

See SentimentDetectionJobProperties

sentiment_score

See SentimentScore

syntax_token

See SyntaxToken

tag

See Tag

topics_detection_job_filter

See TopicsDetectionJobFilter

topics_detection_job_properties

See TopicsDetectionJobProperties

vpc_config

See VpcConfig

Structs

AugmentedManifestsListItem

An augmented manifest file that provides training data for your custom model. An augmented

BatchDetectDominantLanguageItemResult

The result of calling the operation.

BatchDetectEntitiesItemResult

The result of calling the operation. The

BatchDetectKeyPhrasesItemResult

The result of calling the operation. The

BatchDetectSentimentItemResult

The result of calling the operation. The

BatchDetectSyntaxItemResult

The result of calling the operation. The operation

BatchItemError

Describes an error that occurred while processing a document in a batch. The operation

ClassifierEvaluationMetrics

Describes the result metrics for the test data associated with an documentation

ClassifierMetadata

Provides information about a document classifier.

DocumentClass

Specifies the class that categorizes the document being analyzed

DocumentClassificationJobFilter

Provides information for filtering a list of document classification jobs. For more

DocumentClassificationJobProperties

Provides information about a document classification job.

DocumentClassifierFilter

Provides information for filtering a list of document classifiers. You can only specify

DocumentClassifierInputDataConfig

The input properties for training a document classifier.

DocumentClassifierOutputDataConfig

Provides output results configuration parameters for custom classifier jobs.

DocumentClassifierProperties

Provides information about a document classifier.

DocumentLabel

Specifies one of the label or labels that categorize the document being analyzed.

DominantLanguage

Returns the code for the dominant language in the input text and the level of

DominantLanguageDetectionJobFilter

Provides information for filtering a list of dominant language detection jobs. For more

DominantLanguageDetectionJobProperties

Provides information about a dominant language detection job.

EndpointFilter

The filter used to determine which endpoints are returned. You can filter jobs on their

EndpointProperties

Specifies information about the specified endpoint.

EntitiesDetectionJobFilter

Provides information for filtering a list of dominant language detection jobs. For more

EntitiesDetectionJobProperties

Provides information about an entities detection job.

Entity

Provides information about an entity.

EntityLabel

Specifies one of the label or labels that categorize the personally identifiable

EntityRecognizerAnnotations

Describes the annotations associated with a entity recognizer.

EntityRecognizerDocuments

Describes the training documents submitted with an entity recognizer.

EntityRecognizerEntityList

Describes the entity recognizer submitted with an entity recognizer.

EntityRecognizerEvaluationMetrics

Detailed information about the accuracy of an entity recognizer.

EntityRecognizerFilter

Provides information for filtering a list of entity recognizers. You can only specify one

EntityRecognizerInputDataConfig

Specifies the format and location of the input data.

EntityRecognizerMetadata

Detailed information about an entity recognizer.

EntityRecognizerMetadataEntityTypesListItem

Individual item from the list of entity types in the metadata of an entity

EntityRecognizerProperties

Describes information about an entity recognizer.

EntityTypesEvaluationMetrics

Detailed information about the accuracy of an entity recognizer for a specific entity

EntityTypesListItem

An entity type within a labeled training dataset that Amazon Comprehend uses to train a

EventsDetectionJobFilter

Provides information for filtering a list of event detection jobs.

EventsDetectionJobProperties

Provides information about an events detection job.

InputDataConfig

The input properties for a topic detection job.

KeyPhrase

Describes a key noun phrase.

KeyPhrasesDetectionJobFilter

Provides information for filtering a list of dominant language detection jobs. For more

KeyPhrasesDetectionJobProperties

Provides information about a key phrases detection job.

OutputDataConfig

Provides configuration parameters for the output of topic detection jobs.

PartOfSpeechTag

Identifies the part of speech represented by the token and gives the confidence that

PiiEntitiesDetectionJobFilter

Provides information for filtering a list of PII entity detection jobs.

PiiEntitiesDetectionJobProperties

Provides information about a PII entities detection job.

PiiEntity

Provides information about a PII entity.

PiiOutputDataConfig

Provides configuration parameters for the output of PII entity detection jobs.

RedactionConfig

Provides configuration parameters for PII entity redaction.

SentimentDetectionJobFilter

Provides information for filtering a list of dominant language detection jobs. For more

SentimentDetectionJobProperties

Provides information about a sentiment detection job.

SentimentScore

Describes the level of confidence that Amazon Comprehend has in the accuracy of its

SyntaxToken

Represents a work in the input text that was recognized and assigned a part of speech.

Tag

A key-value pair that adds as a metadata to a resource used by Amazon Comprehend. For

TopicsDetectionJobFilter

Provides information for filtering topic detection jobs. For more information, see

TopicsDetectionJobProperties

Provides information about a topic detection job.

VpcConfig

Configuration parameters for an optional private Virtual Private Cloud (VPC) containing

Enums

DocumentClassifierDataFormat

DocumentClassifierMode

EndpointStatus

EntityRecognizerDataFormat

EntityType

InputFormat

JobStatus

LanguageCode

ModelStatus

PartOfSpeechTagType

PiiEntitiesDetectionMaskMode

PiiEntitiesDetectionMode

PiiEntityType

SentimentType

SyntaxLanguageCode