Module aws_sdk_comprehend::operation

Structs

BatchDetectDominantLanguage

Determines the dominant language of the input text for a batch of documents. For a list

BatchDetectEntities

Inspects the text of a batch of documents for named entities and returns information

BatchDetectKeyPhrases

Detects the key noun phrases found in a batch of documents.

BatchDetectSentiment

Inspects a batch of documents and returns an inference of the prevailing sentiment,

BatchDetectSyntax

Inspects the text of a batch of documents for the syntax and part of speech of the words

ClassifyDocument

Creates a new document classification request to analyze a single document in real-time,

ContainsPiiEntities

Analyzes input text for the presence of personally identifiable information (PII) and

CreateDocumentClassifier

Creates a new document classifier that you can use to categorize documents. To create a

CreateEndpoint

Creates a model-specific endpoint for synchronous inference for a previously trained

CreateEntityRecognizer

Creates an entity recognizer using submitted files. After your

DeleteDocumentClassifier

Deletes a previously created document classifier

DeleteEndpoint

Deletes a model-specific endpoint for a previously-trained custom model. All endpoints

DeleteEntityRecognizer

Deletes an entity recognizer.

DescribeDocumentClassificationJob

Gets the properties associated with a document classification job. Use this operation to

DescribeDocumentClassifier

Gets the properties associated with a document classifier.

DescribeDominantLanguageDetectionJob

Gets the properties associated with a dominant language detection job. Use this operation

DescribeEndpoint

Gets the properties associated with a specific endpoint. Use this operation to get the

DescribeEntitiesDetectionJob

Gets the properties associated with an entities detection job. Use this operation to get

DescribeEntityRecognizer

Provides details about an entity recognizer including status, S3 buckets containing

DescribeEventsDetectionJob

Gets the status and details of an events detection job.

DescribeKeyPhrasesDetectionJob

Gets the properties associated with a key phrases detection job. Use this operation to get

DescribePiiEntitiesDetectionJob

Gets the properties associated with a PII entities detection job. For example, you can use

DescribeSentimentDetectionJob

Gets the properties associated with a sentiment detection job. Use this operation to get

DescribeTopicsDetectionJob

Gets the properties associated with a topic detection job. Use this operation to get

DetectDominantLanguage

Determines the dominant language of the input text. For a list of languages that Amazon

DetectEntities

Inspects text for named entities, and returns information about them. For more

DetectKeyPhrases

Detects the key noun phrases found in the text.

DetectPiiEntities

Inspects the input text for entities that contain personally identifiable information

DetectSentiment

Inspects text and returns an inference of the prevailing sentiment

DetectSyntax

Inspects text for syntax and the part of speech of words in the document. For more

ListDocumentClassificationJobs

Gets a list of the documentation classification jobs that you have submitted.

ListDocumentClassifiers

Gets a list of the document classifiers that you have created.

ListDominantLanguageDetectionJobs

Gets a list of the dominant language detection jobs that you have submitted.

ListEndpoints

Gets a list of all existing endpoints that you've created.

ListEntitiesDetectionJobs

Gets a list of the entity detection jobs that you have submitted.

ListEntityRecognizers

Gets a list of the properties of all entity recognizers that you created, including

ListEventsDetectionJobs

Gets a list of the events detection jobs that you have submitted.

ListKeyPhrasesDetectionJobs

Get a list of key phrase detection jobs that you have submitted.

ListPiiEntitiesDetectionJobs

Gets a list of the PII entity detection jobs that you have submitted.

ListSentimentDetectionJobs

Gets a list of sentiment detection jobs that you have submitted.

ListTagsForResource

Lists all tags associated with a given Amazon Comprehend resource.

ListTopicsDetectionJobs

Gets a list of the topic detection jobs that you have submitted.

StartDocumentClassificationJob

Starts an asynchronous document classification job. Use the operation to track the progress of the

StartDominantLanguageDetectionJob

Starts an asynchronous dominant language detection job for a collection of documents. Use

StartEntitiesDetectionJob

Starts an asynchronous entity detection job for a collection of documents. Use the operation to track the status of a job.

StartEventsDetectionJob

Starts an asynchronous event detection job for a collection of documents.

StartKeyPhrasesDetectionJob

Starts an asynchronous key phrase detection job for a collection of documents. Use the

StartPiiEntitiesDetectionJob

Starts an asynchronous PII entity detection job for a collection of documents.

StartSentimentDetectionJob

Starts an asynchronous sentiment detection job for a collection of documents. use the

StartTopicsDetectionJob

Starts an asynchronous topic detection job. Use the

StopDominantLanguageDetectionJob

Stops a dominant language detection job in progress.

StopEntitiesDetectionJob

Stops an entities detection job in progress.

StopEventsDetectionJob

Stops an events detection job in progress.

StopKeyPhrasesDetectionJob

Stops a key phrases detection job in progress.

StopPiiEntitiesDetectionJob

Stops a PII entities detection job in progress.

StopSentimentDetectionJob

Stops a sentiment detection job in progress.

StopTrainingDocumentClassifier

Stops a document classifier training job while in progress.

StopTrainingEntityRecognizer

Stops an entity recognizer training job while in progress.

TagResource

Associates a specific tag with an Amazon Comprehend resource. A tag is a key-value pair

UntagResource

Removes a specific tag associated with an Amazon Comprehend resource.

UpdateEndpoint

Updates information about the specified endpoint.